+++ packages/create-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @shopify/create-app
 
+## 0.31.3
+
+### Patch Changes
+
+- Fix the detection of the package manager when creating an app
+
 ## 0.31.2
 
 ### Patch Changes