+++ packages/create-hydrogen/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @shopify/create-hydrogen
 
+## 0.31.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @shopify/cli-kit@0.31.3
+  - @shopify/cli-hydrogen@0.31.3
+
 ## 0.31.1
 
 ### Patch Changes